integrate package template into iffy
* update express to handle the new template package
* move “go” and “determine” into express as “directives”, they arent
really templates: they are multi-word expressions.
* merge object generator and object builder — fixes bug with
conflicting ids when using sequences and unnamed objects.
* move join into core, its needed by templates and templates dont
include express
* change stopping to support “one time only”
* simply counters to use int, because why not.
* all object creation delayed until runtime start, rtm now returns error
* untested command: print slash

+++ dl/core/cycle.go
@@ -7,7 +7,7 @@ import (
 
 type CycleCounter struct {
 	Name string `if:"id"`
-	Curr float64
+	Curr int
 }
 
 type CycleText struct {
